# apps/api/generated/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Title",
    "description": "Title",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api"
    }
  ],
  "paths": {
...
    "/playback/queue": {
      "get": {
        "tags": [
          "playback"
        ],
        "summary": "Get queue",
        "description": "Get queue",
        "operationId": "getQueue",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetQueueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/playback/queue/add": {
      "post": {
        "tags": [
          "playback"
        ],
        "summary": "Add to queue",
        "description": "Add to queue",
        "operationId": "addToQueue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToQueueRequest"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddToQueueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/playback/queue/delete": {
      "post": {
        "tags": [
          "playback"
        ],
        "summary": "Remove from queue",
        "description": "Remove from queue",
        "operationId": "removeFromQueue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFromQueueRequest"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveFromQueueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/playback/queue/clear": {
      "delete": {
        "tags": [
          "playback"
        ],
        "summary": "Clear queue",
        "description": "Clear queue",
        "operationId": "clearQueue",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearQueueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/playback/queue/move": {
      "put": {
        "tags": [
          "playback"
        ],
        "summary": "Move in queue",
        "description": "Move in queue",
        "operationId": "moveInQueue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveInQueueRequest"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoveInQueueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/playback/play": {
      "post": {
        "tags": [
          "playback"
        ],
        "summary": "Play track",
        "description": "Play track",
        "operationId": "playTrack",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayTrackRequest"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayTrackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/playback/volume": {
      "post": {
        "tags": [
          "playback"
        ],
        "summary": "Set volume",
        "description": "Set volume",
        "operationId": "setVolume",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "boolean"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetVolumeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/playback/quality": {
      "post": {
        "tags": [
          "playback"
        ],
        "summary": "Set quality",
        "description": "Set quality",
        "operationId": "setQuality",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetQualityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/playback/device/current": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Get current device",
        "description": "Get current device",
        "operationId": "getCurrentDevice",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCurrentDeviceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/playback/device/register": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Register device",
        "description": "Register device",
        "operationId": "registerDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceRequest"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDeviceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Track": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "artist",
              "album",
              "track",
              "playlist"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "album": {
            "type": "string"
          },
          "duration": {
            "type": "integer"
          },
          "image": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "playable": {
            "type": "boolean"
          }
        },
        "required": [
          "type",
          "id",
          "name",
          "artist",
          "album",
          "duration",
          "image",
          "provider",
          "playable"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Device": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "volume": {
            "type": "integer"
          },
          "state": {
            "type": "boolean"
          },
          "ws": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "volume",
          "state",
          "ws"
        ]
      },
...
      "AddToQueueResponse": {
        "type": "object",
        "properties": {
          "track": {
            "$ref": "#/components/schemas/Track"
          }
        }
      },
      "GetQueueResponse": {
        "type": "object",
        "properties": {
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Track"
            }
          }
        }
      },
      "MoveInQueueResponse": {
        "type": "object",
        "properties": {
          "track": {
            "$ref": "#/components/schemas/Track"
          }
        }
      },
      "ClearQueueResponse": {
        "type": "object",
        "properties": {
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Track"
            }
          }
        }
      },
...
      "GetCurrentDeviceResponse": {
        "type": "object",
        "properties": {
          "device": {
            "$ref": "#/components/schemas/Device"
          }
        }
      },
      "PlayTrackResponse": {
        "type": "object",
        "properties": {
          "track": {
            "$ref": "#/components/schemas/Track"
          }
        }
      },
...
      "SetVolumeResponse": {
        "type": "object",
        "properties": {
          "volume": {
            "type": "number"
          }
        }
      },
...
      "SetQualityResponse": {
        "type": "object",
        "properties": {
          "quality": {
            "type": "string"
          }
        }
      },
...
      "RemoveFromQueueResponse": {
        "type": "object",
        "properties": {
          "track": {
            "$ref": "#/components/schemas/Track"
          }
        }
      },
...
      "RegisterDeviceResponse": {
        "type": "object",
        "properties": {
          "device": {
            "$ref": "#/components/schemas/Device"
          }
        }
      },
...
      "AddToQueueRequest": {
        "type": "object",
        "properties": {
          "track": {
            "$ref": "#/components/schemas/Track"
          }
        }
      },
      "RemoveFromQueueRequest": {
        "type": "object",
        "properties": {
          "track": {
            "$ref": "#/components/schemas/Track"
          }
        }
      },
      "MoveInQueueRequest": {
        "type": "object",
        "properties": {
          "track": {
            "$ref": "#/components/schemas/Track"
          },
          "position": {
            "type": "number"
          }
        }
      },
...
      "RegisterDeviceRequest": {
        "type": "object",
        "properties": {
          "ws": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "PlayTrackRequest": {
        "type": "object",
        "properties": {
          "track": {
            "$ref": "#/components/schemas/Track"
          },
          "serviceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}